org: use contract names as filenames
Remove a common hurdle in get acquainted with the codebase.

Previously files were named by their dominant operation rather than the
object that they contained. This choice made sense earlier in the
history of this repo since

a) the core domain of most contracts is contained in a single function,
   with other functions being for assistance or administration,

b) in this more exploratory phase the specific contract layout was less
   durable than the labels to specific fragments of business logic (aka
   the "memes")

Now that the architecture is more settled it makes sense to return to
more conventional naming patterns in the case of the core contracts.